Add files via upload
Successfully integrated the map and housing market css files decreasing load times and therefore lowering bandwidth requirements. However we are now severely lacking functionality for pre existing functions for the map. I also somehow managed to hide the UI for the housing market so I'll fix that in the next commit.